add domain movie mapper

+++ core/domain/entities/src/main/java/com/elhady/entities/MovieEntity.kt
@@ -5,7 +5,7 @@ data class MovieEntity(
     val movieName: String,
     val movieType:String = "movie",
     val movieGenreEntities: List<GenreEntity>,
-    val movieRate: Float,
+    val movieRate: Double,
     val movieImage: String,
     val movieYear: String,
 )
